Add hiding logic to FaucetInfo component
