+++ src/helpers/index.ts
@@ -1,1 +1,1 @@
-export * from "./ContractHelper";
+export * from "./PushTokenContractHelper";
